externalize loadTile function into separate module
