+++ app/components/Stats/searchBuiltupAreas.js
@@ -1,11 +1,8 @@
-import * as request from 'superagent'
-import vt from 'vector-tile'
-import Protobuf from 'pbf'
-import turf from 'turf'
-import { extent, intersect, bboxPolygon, featurecollection, centroid, lineDistance, within } from 'turf'
+import { area, extent, intersect, bboxPolygon, featurecollection, centroid, lineDistance, within } from 'turf'
 import Sphericalmercator from 'sphericalmercator'
 import { queue } from 'd3-queue'
-import settings from '../../settings/settings'
+import loadTile from '../Map/loadVectorTile.js'
+import { getRegionZoom, getRegionTiles } from './searchFeatures.js'
 
 var merc = new Sphericalmercator({size: 512})
 
@@ -34,116 +31,21 @@ function fetch(region, callback) {
   })
 }
 
-function getRegionZoom(region) {
-  const maxZoom = 13 // todo: setting "maxZoom"
-  const tileLimit = 12 // todo: setting "tileLimit"
-  const regionBounds = extent(region)
-  for (let z=maxZoom; z>0; z--) {
-    let tileBounds = merc.xyz(regionBounds, z)
-    let tilesNum = (1 + tileBounds.maxX - tileBounds.minX) * (1 + tileBounds.maxY - tileBounds.minY)
-    if (tilesNum <= tileLimit) {
-      return z
-    }
-  }
-  return 0
-}
-
-function getRegionTiles(region, zoom) {
-  const regionBounds = extent(region)
-  var tiles = []
-  // get all tiles for the regions bbox
-  var tileBounds = merc.xyz(regionBounds, zoom)
-  for (let x=tileBounds.minX; x<=tileBounds.maxX; x++) {
-    for (let y=tileBounds.minY; y<=tileBounds.maxY; y++) {
-      tiles.push({
-        x,
-        y,
-        zoom,
-        hash: x+'/'+y+'/'+zoom
-      })
-    }
-  }
-  // drop tiles that are actually outside the region
-
-  tiles = tiles.filter(tile => {
-
-    const bboxPoly = bboxPolygon(merc.bbox(tile.x, tile.y, tile.zoom))
-    try {
-      return intersect(
-        bboxPoly,
-        region
-      )
-    } catch(e) {
-      console.warn(e)
-      return true
-    }
-  }
-  )
-  return tiles
-}
-
 function getAndCacheTile(tile, callback) {
   const cachePage = 'builtup'
-  loadTile(tile, function(err, data) {
+  loadTile('http://129.206.7.145:7778/{z}/{x}/{y}.pbf', 'buildup', tile, function(err, data) {
     if (err) return callback(err)
     // convert features to centroids, store tile data in cache
     data.features = data.features.map(feature => {
       var centr = centroid(feature)
       centr.properties = feature.properties
       centr.properties.tile = tile
-      centr.properties.area = centr.properties.area || turf.area(feature.geometry)
+      centr.properties.area = centr.properties.area || area(feature.geometry)
       return centr
     })
     cache[cachePage][tile.hash] = data
     callback(null) // don't return any actual data as it is available via the cache already
   })
 }
-
-function parseTile(tile, data, callback) {
-  const layer = data.layers['buildup'] // todo: settings?
-  var features = []
-  if (layer) {
-    for (let i=0; i<layer.length; i++) {
-      let feature = layer.feature(i)
-      features.push(feature.toGeoJSON(tile.x, tile.y, tile.zoom))
-    }
-  }
-  callback(null, featurecollection(features))
-}
-function loadTile(tile, callback) {
-  // based on https://github.com/mapbox/mapbox-gl-js/blob/master/js/source/worker.js
-  var url = 'http://129.206.7.145:7778/'+tile.zoom+'/'+tile.x+'/'+tile.y+'.pbf'
-
-  getArrayBuffer(url, function done(err, data) {
-    if (err) return callback(err)
-    if (data === null) return callback(null, featurecollection([]))
-    data = new vt.VectorTile(new Protobuf(new Uint8Array(data)))
-    parseTile(tile, data, callback)
-  })
-}
-function getArrayBuffer(url, callback) {
-  // todo: global?
-  request.parse['application/x-protobuf'] = obj => obj
-  request.parse['application/octet-stream'] = obj => obj
-
-  /* eslint-disable indent */
-  request.get(url)
-  .on('request', function () {
-    // todo: needed?
-    // todo: check browser compat?? xhr2??? see https://github.com/visionmedia/superagent/pull/393 + https://github.com/visionmedia/superagent/pull/566
-    this.xhr.responseType = 'arraybuffer' // or blob
-  })
-  .end(function(err,res) {
-    // now res.body is an arraybuffer or a blob
-    if (!err && res.status >= 200 && res.status < 300) {
-      callback(null, res.body)
-    } else if (res && res.status === 404) {
-      callback(null, null)
-    } else {
-      callback(err || new Error(res.status))
-    }
-  });
-  /* eslint-enable indent */
-};
 
 export default fetch
