+++ app/components/Stats/searchFeatures.js
@@ -1,9 +1,7 @@
-import * as request from 'superagent'
-import vt from 'vector-tile'
-import Protobuf from 'pbf'
-import { extent, intersect, bboxPolygon, featurecollection, centroid, lineDistance, within } from 'turf'
+import { area, extent, intersect, bboxPolygon, featurecollection, centroid, lineDistance, within } from 'turf'
 import Sphericalmercator from 'sphericalmercator'
 import { queue } from 'd3-queue'
+import loadTile from '../Map/loadVectorTile.js'
 import settings from '../../settings/settings'
 
 var merc = new Sphericalmercator({size: 512})
@@ -61,16 +59,14 @@ function getRegionTiles(region, zoom) {
       tiles.push({
         x,
         y,
-        zoom,
+        z: zoom,
         hash: x+'/'+y+'/'+zoom
       })
     }
   }
   // drop tiles that are actually outside the region
-
   tiles = tiles.filter(tile => {
-
-    const bboxPoly = bboxPolygon(merc.bbox(tile.x, tile.y, tile.zoom))
+    const bboxPoly = bboxPolygon(merc.bbox(tile.x, tile.y, tile.z))
     try {
       return intersect(
         bboxPoly,
@@ -80,14 +76,19 @@ function getRegionTiles(region, zoom) {
       console.warn(e)
       return true
     }
-  }
-  )
+  })
   return tiles
 }
 
 function getAndCacheTile(tile, filter, time, callback) {
   const cachePage = (!time || time === 'now') ? filter : time + '/' + filter
-  loadTile(tile, filter, time, function(err, data) {
+  var url
+  if (!time || time === 'now') {
+    url = settings['vt-source']+'/'+filter+'/{z}/{x}/{y}.pbf'
+  } else {
+    url = settings['vt-hist-source']+'/'+time+'/'+filter+'/{z}/{x}/{y}.pbf'
+  }
+  loadTile(url, 'osm', tile, function(err, data) {
     if (err) return callback(err)
     // convert features to centroids, store tile data in cache
     data.features = data.features.map(feature => {
@@ -104,57 +105,5 @@ function getAndCacheTile(tile, filter, time, callback) {
   })
 }
 
-function parseTile(tile, data, callback) {
-  const layer = data.layers['osm'] // todo: settings?
-  var features = []
-  if (layer) {
-    for (let i=0; i<layer.length; i++) {
-      let feature = layer.feature(i)
-      features.push(feature.toGeoJSON(tile.x, tile.y, tile.zoom))
-    }
-  }
-  callback(null, featurecollection(features))
-}
-function loadTile(tile, filter, time, callback) {
-  // based on https://github.com/mapbox/mapbox-gl-js/blob/master/js/source/worker.js
-  var url
-  if (!time || time === 'now') {
-    url = settings['vt-source']+'/'+filter+'/'+tile.zoom+'/'+tile.x+'/'+tile.y+'.pbf'
-  } else {
-    url = settings['vt-hist-source']+'/'+time+'/'+filter+'/'+tile.zoom+'/'+tile.x+'/'+tile.y+'.pbf'
-  }
-
-  getArrayBuffer(url, function done(err, data) {
-    if (err) return callback(err)
-    if (data === null) return callback(null, featurecollection([]))
-    data = new vt.VectorTile(new Protobuf(new Uint8Array(data)))
-    parseTile(tile, data, callback)
-  })
-}
-function getArrayBuffer(url, callback) {
-  // todo: global?
-  request.parse['application/x-protobuf'] = obj => obj
-  request.parse['application/octet-stream'] = obj => obj
-
-  /* eslint-disable indent */
-  request.get(url)
-  .on('request', function () {
-    // todo: needed?
-    // todo: check browser compat?? xhr2??? see https://github.com/visionmedia/superagent/pull/393 + https://github.com/visionmedia/superagent/pull/566
-    this.xhr.responseType = 'arraybuffer' // or blob
-  })
-  .end(function(err,res) {
-    // now res.body is an arraybuffer or a blob
-    if (!err && res.status >= 200 && res.status < 300) {
-      callback(null, res.body)
-    } else if (res && res.status === 404) {
-      callback(null, null)
-    } else {
-      callback(err || new Error(res.status))
-    }
-  });
-  /* eslint-enable indent */
-};
-
 export default fetch
-export { getRegionZoom }
+export { getRegionZoom, getRegionTiles }
